% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C7D10.R
\name{C7D10}
\alias{C7D10}
\title{Seven clusters in ten dimensions having unequal within cluster variations.}
\usage{
C7D10(n)
}
\arguments{
\item{\code{n}}{number of observations in the dataset. These are equally divided between clusters.}
}
\value{
Returns a list having two components described as follows:
\describe{
 \item{data}{dataset generated from a data generating process having \code{n} observations and \code{K} number of clusters}.
\item{truelab}{clustering label vector correspounding to the known data generating process}
}
}
\description{
All the clusters are generated from Gaussain distributions. The dimensions are i.i.d.
}
\details{
each cluster has equal number of observations. The dataset has 10 dimensions and \code{K=7} number of clusters.
}
\examples{
dmat <- C7D10(175)
plot(dmat$data, col = dmat$truelab, pch = 16)
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
