% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filldys.R
\name{filldys}
\alias{filldys}
\title{filldys}
\usage{
filldys(dys)
}
\arguments{
\item{dys}{A vector of pairwise distances between observations. Usually an object of class "dist"}
}
\value{
Returns a vector of pairwise distances between observations. The total entries are \code{n*(n-)/2+1}
}
\description{
An auxilary function for HOSIL, OSIL, FOSIL and PAMSIL.
}
\details{
filldys returns a vector of distances that append 0 in the start. This is to deal with the main diagonal entry of the distance matrix. 0 is appended in the begining to obtain the diagonal values which are all zeros in the distance matrix whenever needed. Note that other alternatives might not work for instance c(0, dys). Do not try to change
}
