% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOSIL.R
\name{FOSIL}
\alias{FOSIL}
\title{Fast optimum avearage silhouette width clustering}
\usage{
FOSIL(dmat, distmethod="euclidean", k=4, v=0.2, many.sample=2)
}
\arguments{
\item{dmat}{Either a numeric matrix or data frame of observed values.
The row represent observations to cluster and columnn represents the variables.}

\item{distmethod}{the distance mehtod to be used. Current available methods are "euclidean", 
"maximum", "manhattan", "canberra", "binary" or "minkowski". 
See \code{\link[stats:dist]{dist}} for more details on these methods.}

\item{k}{maximum value to be used for the estimation of number of clusters.}

\item{v}{data proportion to be used for samplig.}

\item{many.sample}{how many times random samples of size v to be taken from data.}
}
\value{
Returns a list having following components:
\describe{
\item{n}{total number of data points.}
\item{K}{estimated number of clusters.}
\item{clus_lab}{clustering label vector.}
\item{clus_size}{cluster sizes.}
\item{silh}{OASW value of each cluster.}
\item{avg_clus_silh}{OASW for each cluster.}
\item{avg_silh}{OASW value for the clustering.}
\item{iter}{number of iteration taken by the algorithm to converge.}
\item{best_init_method}{method that gave the best initialization i.e., maximum ASW.}
}
}
\description{
This is the fast version of the OSil algorithm. 
OSil is an optimum average silhouette width (OASW) clustering method that 
donot use any kind of cluster centriods. Only data is needed as input. 
The algorithm can estimate number of clusters.
}
\details{
OSIL has an initialization phase. Based on extensive simulaitons the best
 initialization methods from among a wide range of existing clustering methods, 
  for the algorithm has been identified namely average, Ward's, pam,
   kmeans, and model-based clustering. In case distances are provided for 
   clustering kmeans and model-based clustering are excluded from the initialization methods.
}
\examples{
require(mlbench)
dmat <- mlbench.shapes(100)$x
oasw_clus <- FOSIL(dmat) 
plot(dmat, col = oasw_clus$clus_lab, pch = 16, cex = 1.5)
}
\references{
Hartigan, J. A. and Wong, M. A. (1979). Algorithm AS 136: A K-means clustering algorithm. 
\emph{Applied Statistics}, 28, 100–108.

Kaufman, L. and P. J. Rousseeuw (1990). Finding groups in data: 
an introduction to cluster analysis, Volume 344. John Wiley and Sons.

R. R. Sokal and C. D. Michener. A statistical method for evaluating systematic relationships.
 \emph{Univesity Kansas Science Bulletin}, 38(22):1409–1438, 1958.
 
J. H. Ward Jr. Hierarchical grouping to optimize an objective function.
 \emph{Journal of the American Statistical Association}, 58(301):236–244, 1963.
 
C. Fraley and A. E. Raftery. How many clusters? which clustering method?
 answers via model-based cluster analysis. \emph{The Computer Journal}, 41(8):578–588, 1998.
 
Batool F. (2019). Optimum average silhouette width clustering. 
 \emph{PhD Thesis}, University College London.
}
\seealso{
\code{\link{OSIL}} for not so fast version.
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
