% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOSIL.R
\name{HOSIL}
\alias{HOSIL}
\title{Hierarchical optimum average SILhouette width clustering}
\usage{
HOSIL(dys,  distmethod = "euclidean", fix.K = 3)
}
\arguments{
\item{dys}{A vector of pairwise distances between observations. Usually an object of class \code{"dist"} or a data matrix or data frame. In latter case also needs distance method. The default is set at \code{euclidean}.}

\item{distmethod}{the distance mehtod to be used. Current available methods are \code{euclidean}, \code{maximum}, \code{manhattan}, \code{canberra}, \code{binary}, or \code{minkowski}. See \code{\link[stats:dist]{dist}} for more details on these methods.}

\item{fix.K}{user defined number of clusters against which clustering is requried.}

\item{...}{other parameters}
}
\value{
Returns a list with th efollwoing components

\describe{
\item{est_K}{estimated number of clusters }
\item{clus_vect_est}{clustering label vector for the estimated number of clusters}
\item{OASW_est_K}{Value of the objective function against the estimated number of clusters}
\item{all_OASW}{Values of the objective function (OASW linkage) against \code{n-1} to \code{2} number of clusters. Thus the first value \code{all_OASW[1]} gives the OASW linkage value against \code{(n-1)} number of clusters, the next value contains \code{(n-2)} then \code{(n-3),..., 3, 2}}
\item{all_clus}{A vector of length \code{n*(n-2)} containing the clustering label vectors for all number of clusters from \code{(n-1)} to 2, where n is number of observations. So first n values will be clustering labels for n-1 number of clusters the next n will be clustering lables for n-2 clusters and so on. For n number of observations if clustering for K number of clusters is requried do: \cr
 \code{r <- n-K} \cr
 \code{all_clus[(n*(r-1)+1):(n*r)] + 1 }\cr
 
to see clustering labels against that K. For the user convience fix.K is also created to reterive clustering against one of these K. } }
\describe{
If the number of clusters are specified by the user some additional output is also returned as follows
\item{fix_K}{user specified number of clusters }
\item{clus_vect_fix}{clustering label vector for the user specified number of clusters}
\item{OASW_fix_K}{Value of the objective function against the user specified number of clusters}
}
}
\description{
HOSIL is a clustering alogromative hierarchical clustering algorithm.
 The cluster mearge are defined at each hierarchy using a new linkage method defined by
  the optimization of the ASW index. The method can also estimate the number of clusters 
  based on OASW linkage criterion. If number of clsuters is to be fixed
   (see \link{\code{fix.k}}) the minimum allowd is 3 and can be at most \code{(n-1)}.
}
\examples{
require(sn)
data_5 <- function(n){
K <- 6
x2 <- rexp(n/K, rate = 10)
y2 <- rexp(n/K, rate = 10)
shape1 <- 2
shape2 <- 3
x7 <- rbeta(n/K, shape1, shape2, ncp = 220)
y7 <- rbeta(n/K, shape1, shape2, ncp = 120)
shape = 10
x6 <- rweibull(n/K, shape, scale = 4)
y6 <- rweibull(n/K, shape, scale = 4)
rate  <-  2
x4 <- rgamma(n = n/K, shape = 15, rate )
y4 <- rgamma(n = n/K, shape = 15)
x10 <-runif(n/K, -6, -2)
y10 <- runif(n/K, -6, -2)
x11 <-  rsn(n/K, xi = 5, omega = 0.6, alpha = 4, tau = 5)
y11 <-  rsn(n/K,xi = 0, omega = 0.6, alpha = 4, tau = 5)
data <- data.frame(x=c(x2, x7, x6, x4, x11, x10), y=c(y2, y7, y6, y4, y11, y10))
truelab <- rep(c(1, 2, 3, 4, 5, 6), each = n/K)
out <- list(data=data, truelab=truelab)
return(out)
}
dmat <- data_5(96)$data
dys <- dist(dmat)
HOSIL(dys) 
oasw_clus <- HOSIL(dys, fix.K=6)
plot(dmat, col = oasw_clus$clus_vect_fix, pch = 16)
}
\references{
Batool F. (2019). Optimum average silhouette width clustering.  
\emph{PhD Thesis}, University College London.

Kaufman, L. and P. J. Rousseeuw (1990). Finding groups in data: an introduction to cluster analysis,
 Volume 344. John Wiley & Sons.
}
\author{
Fatima Batool \email{ ucakfba@ucl.ac.uk }
}
