% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{initialization function for OSIL and FOSIL.}
\usage{
init(dmat, dys, KK=2, indicator=1)
}
\description{
This function was originally wrote to provide the initialization for OSIL and FOSIL
 but can have standalone usage as well. More discription to follow.
}
