% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiDist.R
\name{MultiDist}
\alias{MultiDist}
\title{Six clusters model}
\usage{
MultiDist(n)
}
\arguments{
\item{n}{number of observations in the dataset. These are equally divided between clusters}
}
\value{
Returns a list having two components described as follows:
\describe{
 \item{data}{dataset generated from a data generating process having n observations and K number of clusters}
\item{truelab}{clustering label vector correspounding to the known data generating process}
}
}
\description{
Must have a description. add some
}
\details{
each cluster has equal number of observations. The dataset has 2 dimensions and K=6 number of clusters
}
\examples{
dmat <- MultiDist(100)
plot(dmat$data, col = dmat$truelab, pch = 16)
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
