% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OASW_fun.R
\name{OASW_fun}
\alias{OASW_fun}
\title{OASW_fun}
\usage{
OASW_fun(dys, n, K, clus_lab)
}
\arguments{
\item{dys}{pairwise distances between observations}

\item{n}{number of observations}

\item{K}{number of clusters}

\item{clus_lab}{initial clustering labels}
}
\value{
Returns a list having following components
\describe{
\item{n}{total number of data points}
\item{K}{estimated number of clusters}
\item{clus_lab}{clustering label vector}
\item{clus_size}{cluster sizes}
\item{silh}{OASW value of each cluster }
\item{avg_clus_silh}{OASW for each cluster}
\item{avg_silh}{OASW value for the clustering}
\item{iter}{number of iteration taken by the algorithm to converge} }
}
\description{
This is a wrapper function for C functions.  This function is called from within OSIL funciton.  Can be called standalone for clustering for fixed number of clusters with an intialization label set. This will return an OASW clsutering based on labels
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
