% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAMSIL.R
\name{PAMSIL}
\alias{PAMSIL}
\title{A new partitioning around medoids algorithm}
\usage{
PAMSIL(dys, K, distmethod = "euclidean")
}
\arguments{
\item{dys}{Either a numeric matrix or data frame of observed values or a vector of pairwise distances between observations. In first case the row represent observations to cluster and columnn represents the variables. If data matrix is provided needs to specify the distance method as well. In second case usually an object of class \code{"dist"}. Missing values are not allowed in both cases.}

\item{K}{number of clusters}

\item{distmethod}{distance method to be used}
}
\value{
Returns a list having following components:
\describe{
\item{clus_vect}{clustering label vector}
\item{silh}{optimum ASW value for each data point in the clustering}
\item{clus_size}{cluster sizes}
\item{index_by_cluster}{represents the data index in each cluster. The order is same as that in \code{clus_vect}}
\item{avg_clus_silh}{ASW for each cluster}
\item{avg_silh}{OASW value for the clustering}
\item{iter}{number of iteration taken by the algorithm to converge}
}
}
\description{
An OASW clustering method based on medoids. 
A PAM like clustering algorithm for the optimization of ASW based on medoids proposed in Van et al. (2003). see
}
\details{
This function is based on the stand alone \code{C} functions written by Van et al. (2003).
}
\examples{
dmat <- data(iris)
dys <- dist(dmat)
PAMSIL(dys, 3) 
PAMSIL(dmat, 3, distmethod = "manhattan")

}
\references{
Van der Laan, M., Pollard, K., & Bryan, J. (2003). A new partitioning around medoids
 algorithm. \emph{Journal of Statistical Computation and Simulation}, 73(8), 575-584.
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
