% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoGaussian.R
\name{TwoGaussian}
\alias{TwoGaussian}
\title{TwoGaussian}
\usage{
TwoGaussian(n=100)
}
\arguments{
\item{\code{n}}{number of observations in the dataset. These are equally divided between clusters}
}
\value{
Returns a list having two components described as follows:
\describe{
 \item{data}{dataset generated from a data generating process having \code{n} observations and \code{K=2} number of clusters}
\item{truelab}{clustering label vector correspounding to the known data generating process}
}
}
\description{
Two Gaussian clusters of unequall within cluster variations.
}
\details{
each cluster has equal number of observations. The dataset has 2 dimensions and K=2 number of clusters
}
\examples{
dmat <- TwoGaussian(100)
plot(dmat$data, col = dmat$truelab, pch = 16)
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
